+++ core/admin/panel/Settings.php
@@ -25,7 +25,7 @@ class Settings extends PanelSettings {
 					'icon'  => 'ui-ban',
 					'info'  => __( 'Some of the components can use these exceptions to limit the logging and skip some events.', 'coreactivity' ),
 				),
-				'geo'    => array(
+				'geo'           => array(
 					'title' => __( 'Geo Location', 'coreactivity' ),
 					'icon'  => 'ui-globe',
 					'info'  => __( 'Plugin can use online geolocation services, or the local database for geolocation.', 'coreactivity' ),