+++ core/build.php
@@ -1,1 +1,1 @@
-<?php $build = 1200;
+<?php $build = 1300;